pgd/requirements.txt
# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Web Frameworks
flask>=2.0.0
flask-cors>=3.0.10
fastapi>=0.68.0
uvicorn>=0.15.0

# Image Processing
Pillow>=8.3.0

# Data Processing & Visualization
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
python-multipart>=0.0.5 
requests>=2.26.0

# Security
python-dotenv>=0.19.0  # 환경 변수 관리
pydantic>=1.8.0       # 데이터 검증

# Development Tools
pytest>=6.2.0
black>=21.6b0
flake8>=3.9.0